# server/docs/swagger.yaml
basePath: /
definitions:
...
  apperr.Error:
    properties:
      code:
        type: string
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  url.URLPreviewResponse:
    properties:
      domain:
        type: string
      is_active:
        type: boolean
      is_https:
        type: boolean
      original_url:
        type: string
      safety_rating:
        description: SAFE, SUSPICIOUS
        type: string
      short_code:
        type: string
      short_url:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/semmidev/url-shortener
  description: High-Performance Modular Monolith URL Shortener REST API in Go.
  license:
    name: MIT
    url: https://github.com/semmidev/url-shortener/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: URL Shortener API
  version: "1.0"
paths:
...
  /{code}/preview:
    get:
      parameters:
      - description: Short Code
        in: path
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/url.URLPreviewResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      summary: Preview short URL destination and safety analysis
      tags:
      - Redirect
  /{code}/qr:
    get:
      parameters:
      - description: Short Code
        in: path
        name: code
        required: true
        type: string
      responses:
        "200":
          description: PNG Image
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperr.Error'
      summary: Get QR Code image (PNG) for short code
      tags:
      - Redirect
...
securityDefinitions:
  BearerAuth:
    description: Input your Bearer token in the format "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"